biomodels_qc/requirements.txt
biosimulators_utils[sbml,containers] >= 0.1.64
cement
kisao
lxml
nbformat
owlready2
pypdf2
python_copasi
python_dateutil
python_libsbml
scipy
termcolor
xmldiff
